# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Módulo 1 - Historia Clínica Electrónica (HCE)",
    "description": "Repositorio central de datos médicos del sistema Health Grid. Gestiona consultas, diagnósticos, tratamientos y fichas médicas.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/patients/{id_paciente}/episodes/{id_episodio}/evoluciones": {
      "get": {
        "tags": [
          "Atención Clínica — Evoluciones"
        ],
        "summary": "Listar evoluciones médicas de un episodio",
        "description": "Lista todas las notas de evolución registradas dentro de un episodio, ordenadas cronológicamente.",
        "operationId": "listar_evoluciones_api_v1_patients__id_paciente__episodes__id_episodio__evoluciones_get",
        "security": [
          {
            "JWT Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_paciente",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Paciente"
            }
          },
          {
            "name": "id_episodio",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Episodio"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvolucionListResponse"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Atención Clínica — Evoluciones"
        ],
        "summary": "Registrar una evolución médica en un episodio",
        "description": "Permite registrar una nota de evolución clínica dentro de un episodio activo. El profesional autor se obtiene automáticamente del token JWT.",
        "operationId": "crear_evolucion_api_v1_patients__id_paciente__episodes__id_episodio__evoluciones_post",
        "security": [
          {
            "JWT Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_paciente",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Paciente"
            }
          },
          {
            "name": "id_episodio",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Episodio"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvolucionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvolucionSchema"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
    "/api/v1/patients/{id_paciente}/episodes/{id_episodio}/evoluciones/{id_evolucion}": {
      "get": {
        "tags": [
          "Atención Clínica — Evoluciones"
        ],
        "summary": "Obtener detalle de una evolución médica",
        "description": "Retorna el contenido completo de una nota de evolución específica.",
        "operationId": "obtener_evolucion_api_v1_patients__id_paciente__episodes__id_episodio__evoluciones__id_evolucion__get",
        "security": [
          {
            "JWT Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_paciente",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Paciente"
            }
          },
          {
            "name": "id_episodio",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Episodio"
            }
          },
          {
            "name": "id_evolucion",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Evolucion"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvolucionSchema"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Atención Clínica — Evoluciones"
        ],
        "summary": "Actualizar parcialmente una evolución médica",
        "description": "Permite corregir el contenido de una nota de evolución existente.",
        "operationId": "actualizar_evolucion_api_v1_patients__id_paciente__episodes__id_episodio__evoluciones__id_evolucion__patch",
        "security": [
          {
            "JWT Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "id_paciente",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Paciente"
            }
          },
          {
            "name": "id_episodio",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Episodio"
            }
          },
          {
            "name": "id_evolucion",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Id Evolucion"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvolucionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvolucionSchema"
                }
              }
            }
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/hce/notify-permission-change": {
      "post": {
        "tags": [
          "Integración M10 (Core)"
        ],
        "summary": "Notificar a Core un cambio de permisos originado en HCE",
        "description": "Proxy que reenvía la notificación al Módulo 10 (Core).",
        "operationId": "notify_permission_change_api_v1_hce_notify_permission_change_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PermissionChangeNotification"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionChangeResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Entity",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT Bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "examples": [
              "NOT_FOUND"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "examples": [
              "El recurso solicitado no fue encontrado."
            ]
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse",
        "description": "Respuesta estándar de error del módulo HCE."
      },
...
      "EvolucionCreate": {
        "properties": {
          "contenido": {
            "type": "string",
            "title": "Contenido",
            "description": "Texto libre con la nota clínica del profesional.",
            "examples": [
              "Paciente refiere mejoría del cuadro respiratorio. Se auscultan campos pulmonares limpios. Se indica continuar tratamiento."
            ]
          },
          "fecha": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha",
            "description": "Fecha/hora de la evolución. Si no se envía, se usa la fecha/hora actual."
          }
        },
        "type": "object",
        "required": [
          "contenido"
        ],
        "title": "EvolucionCreate",
        "description": "Body para registrar una nueva evolución médica en un episodio."
      },
      "EvolucionListResponse": {
        "properties": {
          "id_episodio": {
            "type": "integer",
            "title": "Id Episodio"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "evoluciones": {
            "items": {
              "$ref": "#/components/schemas/EvolucionSchema"
            },
            "type": "array",
            "title": "Evoluciones"
          }
        },
        "type": "object",
        "required": [
          "id_episodio",
          "total",
          "evoluciones"
        ],
        "title": "EvolucionListResponse",
        "description": "Respuesta de listado de evoluciones de un episodio."
      },
      "EvolucionSchema": {
        "properties": {
          "id_evolucion": {
            "type": "integer",
            "title": "Id Evolucion"
          },
          "id_episodio": {
            "type": "integer",
            "title": "Id Episodio"
          },
          "id_profesional": {
            "type": "integer",
            "title": "Id Profesional"
          },
          "contenido": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contenido"
          },
          "fecha": {
            "type": "string",
            "format": "date-time",
            "title": "Fecha"
          }
        },
        "type": "object",
        "required": [
          "id_evolucion",
          "id_episodio",
          "id_profesional",
          "fecha"
        ],
        "title": "EvolucionSchema",
        "description": "Schema de respuesta completo de una evolución médica."
      },
      "EvolucionUpdate": {
        "properties": {
          "contenido": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contenido",
            "description": "Texto libre con la nota clínica corregida."
          }
        },
        "type": "object",
        "title": "EvolucionUpdate",
        "description": "Body para actualizar parcialmente una evolución médica (PATCH)."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PermissionChangeNotification": {
        "properties": {
          "id_usuario_afectado": {
            "type": "integer",
            "title": "Id Usuario Afectado",
            "examples": [
              42
            ]
          },
          "tipo_cambio": {
            "$ref": "#/components/schemas/TipoCambioPermiso",
            "examples": [
              "revoke-access"
            ]
          },
          "recurso_afectado": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recurso Afectado",
            "examples": [
              "hce:write"
            ]
          },
          "motivo": {
            "type": "string",
            "title": "Motivo",
            "examples": [
              "Profesional suspendido por auditoría interna."
            ]
          },
          "id_usuario_notificador": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id Usuario Notificador",
            "examples": [
              10
            ]
          },
          "fecha_ocurrencia": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fecha Ocurrencia"
          }
        },
        "type": "object",
        "required": [
          "id_usuario_afectado",
          "tipo_cambio",
          "motivo"
        ],
        "title": "PermissionChangeNotification",
        "description": "Payload que HCE envía a Core para notificar cambio de permisos."
      },
      "PermissionChangeResponse": {
        "properties": {
          "acknowledged": {
            "type": "boolean",
            "title": "Acknowledged",
            "examples": [
              true
            ]
          }
        },
        "type": "object",
        "required": [
          "acknowledged"
        ],
        "title": "PermissionChangeResponse",
        "description": "Respuesta de Core tras recibir la notificación."
      },
...
      "TipoCambioPermiso": {
        "type": "string",
        "enum": [
          "revoke-access",
          "role-change",
          "deactivate-user"
        ],
        "title": "TipoCambioPermiso"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "JWT Bearer": {
        "type": "http",
        "description": "JWT emitido por Módulo 10 (Core). Header: Authorization: Bearer <token>",
        "scheme": "bearer"
      }
    }
  }
}